Append Spotify data to albums that are missing it.
